+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "client",
-  "version": "0.15.2",
+  "version": "0.16.0",
   "private": true,
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^6.3.0",
