+++ src/components/Error.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Error({ children }) {
+  return <p>{children}</p>;
+}
