+++ src/components/Loading.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Loding() {
+  return <p>로딩중입니다...</p>;
+}
