+++ src/index.jsx
@@ -3,15 +3,20 @@ import ReactDOM from 'react-dom/client';
 import { RouterProvider } from 'react-router-dom';
 import { RecoilRoot } from 'recoil';
 
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import router from './routes';
 import GlobalStyles from './GlobalStyles';
+
+const queryClient = new QueryClient();
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <RecoilRoot>
-      <GlobalStyles />
-      <RouterProvider router={router} />
+      <QueryClientProvider client={queryClient}>
+        <GlobalStyles />
+        <RouterProvider router={router} />
+      </QueryClientProvider>
     </RecoilRoot>
   </React.StrictMode>,
 );
