--- src/recoil/loginUser/atom.js
@@ -1,8 +1,0 @@
-import { atom } from 'recoil';
-
-const loginUserAtom = atom({
-  key: 'loginUser',
-  default: null,
-});
-
-export default loginUserAtom;
