+++ src/recoil/userId/atom.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const userIdAtom = atom({
+  key: 'userId',
+  default: null,
+});
+
+export default userIdAtom;
